refactor: rename agents files to on-prem-agents
Align endpoint, test, and mock paths with OnPremAgents naming
after the type/accessor rename. Updates imports and describe blocks.
