+++ test/mocks/on-prem-agents/delete.json
