+++ test/on-prem-agents.integration.test.ts
@@ -14,7 +14,7 @@ const integrationReady = Boolean(MAKE_API_KEY && MAKE_ZONE && MAKE_ORGANIZATION)
 /** Valid UUID that is unlikely to exist in the test org */
 const NON_EXISTENT_AGENT_ID = '00000000-0000-4000-8000-000000000000';
 
-(integrationReady ? describe : describe.skip)('Integration: Agents (on-prem)', () => {
+(integrationReady ? describe : describe.skip)('Integration: OnPremAgents', () => {
     const make = new Make(MAKE_API_KEY, MAKE_ZONE);
 
     let agentId: string;
